Fix Segment Offsets and Match Common Segments
Several segments were split differently across overlays. This gives most
identical segments consistent names, fixes several offset segment
mismatches in splat config, and adds new segments where necessary.

I wrote a linter for matching segments against all overlays. It checked
for name similarity, presence, and correct offsets. These were the easy
changes it found.
