+++ config/splat.us.stcen.yaml
@@ -75,8 +75,7 @@ segments:
       - [0x69EC, .data, tile_data] # tile data
       - [0x81EC, .data, tile_data] # tile definitions
       - [0xC60C, .data, gen/sprites]
-      - [0xD414, .rodata, e_init]
-      - [0xD42C, .rodata, cutscene]
+      - [0xD414, .rodata, cutscene]
       - [0xD4B0, .rodata, F890]
       - [0xD4D8, .rodata, e_red_door] # EntityRedDoor
       - [0xD540, .rodata, e_collect] # EntityPrizeDrop
@@ -96,9 +95,9 @@ segments:
       - [0x135B4, c, e_red_door]
       - [0x14264, c, st_common]
       - [0x15714, c, e_collect]
-      - [0x16C90, c, blit_char]
+      - [0x16C80, c, blit_char]
       - [0x16F30, c, e_misc]
-      - [0x1A5F0, c, e_stage_name]
+      - [0x1A420, c, e_stage_name]
       - [0x1B274, c, e_particles]
       - [0x1BAB8, c, e_room_fg]
       - [0x1BBA4, c, popup]
