+++ config/splat.us.stdre.yaml
@@ -60,6 +60,8 @@ segments:
       - [0x1270, .data, e_misc]
       - [0x12D4, .data, e_misc]
       - [0x138C, data]
+      - [0x138C, .data, e_particles]
+      - [0x140C, .data, e_room_fg]
       - [0x1498, .data, gen/rooms]
       - [0x14AC, data]
       - [0x1608, .data, gen/e_layout] # layout entries data
@@ -104,7 +106,7 @@ segments:
       - [0x119C4, .rodata, e_collect]
       - [0x119DC, .rodata, e_misc]
       - [0x11A14, .rodata, e_misc]
-      - [0x11A28, .rodata, 20734] # EntityStageNamePopup
+      - [0x11A28, .rodata, e_stage_name]
       - [0x11A5C, .rodata, prim_helpers]
       - [0x11A64, c, st_debug]
       - [0x11D00, c, e_breakable]
@@ -121,7 +123,10 @@ segments:
       - [0x1B858, c, e_collect]
       - [0x1CDC4, c, blit_char]
       - [0x1D074, c, e_misc]
-      - [0x20734, c]
+      - [0x20564, c, e_stage_name]
+      - [0x213b8, c, e_particles]
+      - [0x21BFC, c, e_room_fg]
+      - [0x21CE8, c, popup]
       - [0x22018, c, prim_helpers]
       - [0x22848, c]
       - [0x23264, .bss, create_entity]
