+++ config/splat.us.stmad.yaml
@@ -51,7 +51,8 @@ segments:
       - [0xBEC, .data, e_collect]
       - [0xED8, .data, e_collect] # g_collectVelocity
       - [0xEF8, .data, e_misc]
-      - [0x1068, data]
+      - [0x1068, .data, e_particles]
+      - [0x10E8, data]
       - [0x1130, data, g_gen_rooms.c]
       - [0x11D4, data] # assets/st/mad/0.objlayout
       # - [0x11E8, data] # assets/st/mad/1.objlayout
@@ -97,6 +98,9 @@ segments:
       - [0x1344C, c, e_collect]
       - [0x149C0, c, e_misc]
       - [0x17B94, c]
+      - [0x182BC, c, e_particles]
+      - [0x18B00, c, e_room_fg]
+      - [0x18BC8, c, popup]
       - [0x18FA0, c, prim_helpers]
       - [0x197D8, .bss, create_entity]
       - [0x197ED, .bss, bss]
