+++ config/splat.us.stno3.yaml
@@ -52,7 +52,8 @@ segments:
       - [0x1140, .data, e_background_bushes_trees]
       - [0x123C, .data, e_trapdoor]
       - [0x124C, .data, e_jewel_sword_puzzle]
-      - [0x1358, .data, e_sky_entities]
+      - [0x1358, .data, e_death_sky_swirl]
+      - [0x1390, .data, e_sky_entities]
       - [0x13A8, data]
       - [0x1ACC, .data, cutscene]
       - [0x1AD4, data]
@@ -65,7 +66,8 @@ segments:
       - [0x20C4, .data, st_common]
       - [0x22C4, .data, e_collect]
       - [0x25B4, .data, e_misc]
-      - [0x26D0, data]
+      - [0x26D0, .data, e_particles]
+      - [0x2750, data]
       - [0x3180, data]
       - [0x3858, .data, water_effects]
       - [0x3914, data]
@@ -149,14 +151,14 @@ segments:
       - [0x37600, .rodata, e_collect] # EntityEquipItemDrop
       - [0x37618, .rodata, e_misc]
       - [0x37650, .rodata, e_misc]
-      - [0x37664, .rodata, 4AFF0] # EntityStageNamePopup
+      - [0x37664, .rodata, e_stage_name] # EntityStageNamePopup
       - [0x37698, .rodata, e_fire_warg] # EntityFireWarg
       - [0x376D0, .rodata, e_fire_warg] # EntityFireWargWaveAttack
       - [0x376E8, .rodata, e_fire_warg] # EntityUnkId2F
       - [0x376FC, .rodata, 4F58C] # EntityWarg RECT
       - [0x37704, .rodata, 4F58C] # EntityWarg
       - [0x37738, .rodata, e_merman2]
-      - [0x3775C, .rodata, 53F38]
+      - [0x3775C, .rodata, e_merman]
       - [0x37784, .rodata, e_bone_scimitar]
       - [0x377A4, .rodata, e_bat] # EntityBat
       - [0x377B8, .rodata, 564B0] # EntityZombie
@@ -166,13 +168,15 @@ segments:
       - [0x37B98, c, entrance_stage_entities]
       - [0x38564, c, e_castle_door]
       - [0x38A74, c, e_background_bushes_trees]
+
       - [0x39198, c, e_transparent_water]
       - [0x394F0, c, e_cavern_door]
       - [0x39C44, c, entrance_weights]
       - [0x3A348, c, e_trapdoor]
       - [0x3A4C0, c, e_jewel_sword_puzzle]
       - [0x3AEF0, c, e_stairway]
-      - [0x3B548, c, e_sky_entities]
+      - [0x3B548, c, e_death_sky_swirl]
+      - [0x3BDC0, c, e_sky_entities]
       - [0x3BF38, c, e_heartroom]
       - [0x3C3FC, c, e_tilemap_shuffler]
       - [0x3C4EC, c, e_outdoor_ents]
@@ -184,15 +188,18 @@ segments:
       - [0x4343C, c, create_entity]
       - [0x43FB4, c, e_red_door]
       - [0x44C64, c, st_common]
-      - [0x45BC0, c, e_collect]
+      - [0x46114, c, e_collect]
       - [0x47680, c, blit_char]
       - [0x47930, c, e_misc]
-      - [0x4AFF0, c]
+      - [0x4AE20, c, e_stage_name]
+      - [0x4BC74, c, e_particles]
+      - [0x4C4B8, c, e_room_fg]
       - [0x4C5A4, c, e_fire_warg]
       - [0x4F58C, c]
+      - [0x506FC, c, popup]
       - [0x50A2C, c, water_effects]
       - [0x52D40, c, e_merman2]
-      - [0x53F38, c]
+      - [0x54B20, c, e_merman]
       - [0x559d0, c, e_bone_scimitar]
       - [0x56264, c, e_bat]
       - [0x564B0, c]
