+++ config/splat.us.stnp3.yaml
@@ -150,7 +150,7 @@ segments:
       - [0x31FD8, .rodata, e_misc] # EntityMessageBox
       - [0x31FEC, .rodata, e_stage_name] # EntityStageNamePopup
       - [0x32020, .rodata, e_merman2] # EntityMerman2
-      - [0x32044, .rodata, 48238] # EntityMerman
+      - [0x32044, .rodata, e_merman] # EntityMerman
       - [0x3206C, .rodata, e_bone_scimitar] # EntityBoneScimitar data
       - [0x3208C, .rodata, e_bat] # EntityBat
       - [0x320A0, .rodata, 49BC8] # EntityZombie
@@ -190,13 +190,13 @@ segments:
       - [0x3D984, c, e_collect]
       - [0x3EEF0, c, blit_char]
       - [0x3F1A0, c, e_misc]
-      - [0x42860, c, e_stage_name]
+      - [0x42690, c, e_stage_name]
       - [0x434E4, c, e_particles]
       - [0x43D28, c, e_room_fg]
       - [0x43E14, c, popup]
       - [0x44144, c, water_effects]
       - [0x46458, c, e_merman2] # Merman Entities
-      - [0x48238, c] # Merman Entities (2)
+      - [0x48238, c, e_merman]
       - [0x490E8, c, e_bone_scimitar] # EntityBoneScimitar
       - [0x4997C, c, e_bat]
       - [0x49BC8, c]
