+++ config/splat.us.stnz0.yaml
@@ -169,7 +169,7 @@ segments:
       - [0x3B0C4, c, create_entity]
       - [0x3BC3C, c, e_red_door]
       - [0x3C8EC, c, st_common]
-      - [0x3D848, c, e_collect]
+      - [0x3DD9C, c, e_collect]
       - [0x3F308, c, blit_char]
       - [0x3F5B8, c, e_misc]
       - [0x42AA8, c, e_particles]
