+++ src/st/dre/e_room_fg.c
@@ -1,0 +1,2 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../e_room_fg.h"
