+++ src/st/no3/e_particles.c
@@ -1,0 +1,2 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../e_particles.h"
