+++ src/st/no3/popup.c
@@ -1,0 +1,2 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../popup.h"
